feat(migrations): creates initial migrations
This commit creates initial migrations for the
applications.

I have added the migrations to create the users
table and the partners table which serve as the
top most table.

Migrations for slackIntegrations, frecklyIntegrations,
and emailIntegrations tables have been added which
have a many to one relationship with both the users
and the partners table.

I have also added a .sequelizerc file to help with
utilizing sequelize cli commands(pointing out where
files reside).

I have also added node modules to the .gitignore file.

[Finishes issue #1]